添加商品规格方法2

+++ base-common/src/main/java/com/henry/shop/common/base/model/dataobj/com/ComVariantGroup.java
@@ -17,7 +17,7 @@ public class ComVariantGroup implements Serializable {
     /**
      * 规格项数量
      */
-    private Long variantNumber;
+    private Integer variantNumber;
     /**
      * 规格组名称
      */
